requirements.txt
opencv-python
numpy
torch
ultralytics
pyserial
mysql-connector-python
python-dotenv
PyYAML
paho-mqtt